added acordion style checklist to the aside
Added the acordion style checklist to the aside with 4 buttons underneath, there are 3 accordions with there respective checkboxes inside. After this there are 4 buttons, 2 for the random features, the third one will ahve text dynamically appended to it based on conditions and a button that will link this page to the login and review page.