added fetch if top100 isnt already in localStorage
Added Conditional that states that if top100 isnt in local storage, it will run a fetch fucntion to retrieve the top 100 movies and save them to localStorage. To avoid the initial load of the webpage to cause errors upon trying to read the movie element while asyncly running the fetch, I created an event. This event will only run ifA- movies does not return as null or B- once the fetch function reaches its final then method