added functions to display movie cards
Added two functions and a starting point. The first function is called displayMovieCards which will take in a movie and dynamically append html to the movie section. The second function called createMovieCards will retrieve an array of movies, and use displayMovieCards based on this array. The starting point for my code will use the custom event I made to ensure the code doesnt run unless the movies are in localStorage