added variables, and functions needed to display movies randomly
Added function called randmizeSelection which returns a random number if the random number hasnt been selected already. Added a randomBasedOnFilter which iterates through the movie list and uses a diff method (from jeremy on stackOverflow) to retrieve movies that match the genres. If no genre selected, we display 5 random movies. Also added an event listener to the checkboxes to append the checked items to an array.