added randomize by weather feature
added three function that are used for the randomize by weather feature. We use open weather api to get the weather information. To get this information we use the location entered, use the retrieveWeatherId function to get the id, then run create card based on the randomizeBasedOnWeather which essentially returns a list of genres based on weather